Move endpoint to get notifications to dedicated controller
