+++ src/main/java/com/rebalance/controllers/UserController.java
@@ -47,14 +47,6 @@ public class UserController {
                         .map(groupMapper::groupToResponse).collect(Collectors.toList()));
     }
 
-    //TODO: move to notification controller
-    @GetMapping("/{id}/notifications")
-    public ResponseEntity<List<NotificationResponse>> getAllNotificationsByUserId(@PathVariable(value = "id") Long userId) {
-        return ResponseEntity.ok(
-                notificationService.findAllByUserIdAndDeleteThem(userId).stream()
-                        .map(notificationMapper::notificationToResponse).collect(Collectors.toList()));
-    }
-
     //TODO: move to auth controller
     @PostMapping()
     public ResponseEntity<UserResponse> register(@RequestBody UserCreateRequest user) {
